backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
alembic==1.13.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator>=2.0,<3
python-multipart>=0.0.9,<1
